+++ packages/terra-form-input/src/Input.jsx
@@ -165,6 +165,7 @@ class Input extends React.Component {
       attributes.defaultValue = defaultValue;
     }
 
+    // showMpageFusionPlaceholder` prop is added for fusion pass through passivity and should be removed after Fusion Phase2 release.
     if (attributes.placeholder && !attributes.showMpageFusionPlaceholder) {
       attributes.placeholder = null;
     }