+++ packages/terra-form-input/tests/jest/InputField.test.jsx
@@ -21,8 +21,8 @@ describe('InputField', () => {
     expect(wrapper).toMatchSnapshot();
   });
 
-  it('should render a InputField with Placeholder', () => {
-    const textarea = <InputField inputId="test-input" label="label" inputAttrs={{ placeholder: 'Enter The Text', showMpageFusionPlaceholder: true }} />;
+  it('should render placeholder when showMpageFusionPlaceholder is passed', () => {
+    const textarea = <InputField inputId="test-input" label="label" inputAttrs={{ placeholder: 'Enter The Text' }} />;
     const wrapper = shallow(textarea);
     const placeholderInput = (wrapper.find('Field')).find('Input#test-input');
     expect(placeholderInput.props()).toHaveProperty('placeholder', 'Enter The Text');